* commit 'f2645be163c8aec74545878aeef86c4354864c93':
  Tests also require Boost
  Enable `-Werror` in CI
  Enable more non-standard warnings
  Fix `-Wold-style-cast`
  Fix `-Wdeprecated`
  Fix `-Wreturn-type`
  Fix `-Wdouble-promotion`
  Fix `-Wuseless-cast`
  Fix `-Wshadow`
  Fix `-Wsign-compare`
  Fix `-Wignored-qualifiers`
  Fix `-Wunused-but-set-variable`
  Fix `-Wunused-parameter`

+++ include/ConsensusCore/Matrix/SparseVector-inl.hpp
@@ -198,6 +198,6 @@ inline void SparseVector::CheckInvariants() const
     assert(0 <= allocatedBeginRow_ && allocatedBeginRow_ < logicalLength_);
     assert(0 <= allocatedEndRow_ && allocatedEndRow_ <= logicalLength_);
     assert(allocatedBeginRow_ <= allocatedEndRow_);
-    assert((allocatedEndRow_ - allocatedBeginRow_) <= (signed)storage_->size());
+    assert((allocatedEndRow_ - allocatedBeginRow_) <= static_cast<int>(storage_->size()));
 }
 }
